requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=0.11.0

# Image Processing
Pillow>=10.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Cryptography (RSA + AES encryption)
pycryptodome>=3.19.0

# Error Correction Coding
reedsolo>=1.7.0

# Visualization & Plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Data Handling
pandas>=2.0.0

# Configuration Management (optional)
pyyaml>=6.0.0

# Development Tools (optional)
ipython>=8.12.0
jupyter>=1.0.0

# Testing (optional)
pytest>=7.3.0

# Image Analysis Tools (for debugging)
scikit-image>=0.21.0